Close jailsonsf/myList#2

+++ api-mylist/api.py
@@ -49,57 +49,62 @@ def create_user():
 
 @app.route('/users/<id_user>', methods=['POST'])
 def update_user(id_user):
-    # query for check username 
-    query_db = db.read('users', 'id', id_user)
-    user = query_db[0] if len(query_db) == 1 else None 
-    if(user == None):
-        return jsonify(error(404, 'User not exists'))
-    
-    else:
-        json = request.get_json()
-
-        username = json['username']
-        name = json['name']
-        email = json['email']
-        date = user['date_creation']
-
-        if (user['username'] != username and db.verify_register('users', 'username', username)):
-            return jsonify(error(400, 'User invalid or user already exists'))
-
+    if request.method == 'POST':
+        # query for check username 
+        query_db = db.read('users', 'id', id_user)
+        user = query_db[0] if len(query_db) == 1 else None 
+        if(user == None):
+            return jsonify(error(404, 'User not exists'))
+        
         else:
-            user = User(id_user, username, name, email, date)
-            
-            if(db.update('users', id_user, user.to_dict())):
-                return jsonify(user.to_dict())
-            
+            json = request.get_json()
+
+            username = json['username']
+            name = json['name']
+            email = json['email']
+            date = user['date_creation']
+
+            if (user['username'] != username and db.verify_register('users', 'username', username)):
+                return jsonify(error(400, 'User invalid or user already exists'))
+
             else:
-                return jsonify(error(203, 'Not Allowed'))
+                user = User(id_user, username, name, email, date)
+                
+                if(db.update('users', id_user, user.to_dict())):
+                    return jsonify(user.to_dict())
+                
+                else:
+                    return jsonify(error(203, 'Not Allowed'))
 
 @app.route('/users/delete/<id_user>', methods=['DELETE'])
 def delete_user(id_user):
-    # query for check username 
-    query_db = db.read('users', 'id', id_user)
-    user = query_db[0] if len(query_db) == 1 else None 
-    if (user == None):
-        return jsonify(error(404, 'User not exists'))
-    
-    else:
-
-        if (db.delete('users', id_user)):
-            
-            # delete all lists
-            lists = db.read('lists', 'id_user', id_user)
-
-            if (len(lists) > 0):
-                for list_user in lists:
-                    print(list_user['id'])
-                    db.delete('lists', list_user['id'])
-
-            return jsonify(error(200, "Removed"))
-
+    if request.method == 'DELETE':
+        # query for check username 
+        query_db = db.read('users', 'id', id_user)
+        user = query_db[0] if len(query_db) == 1 else None 
+        print("ola")
+        if (user == None):
+            return jsonify(error(404, 'User not exists'))
+        
         else:
-            return jsonify(error(400, "Not removed"))    
 
+            if (db.delete('users', id_user)):
+                
+                # delete all lists
+                lists = db.read('lists', 'id_user', id_user)
+
+                if (len(lists) > 0):
+                    for list_user in lists:
+                        print(list_user['id'])
+                        db.delete('lists', list_user['id'])
+
+                return jsonify(error(200, "Removed"))
+
+            else:
+                return jsonify(error(400, "Not removed")) 
+    
+
+# routes list 
 @app.route('/users/lists/<id_user>', methods=['GET'])
 def get_lists_by_user(id_user):
     return jsonify(db.read('lists', 'id_user', id_user))
@@ -116,7 +121,9 @@ def create_list():
     title = json['title']
     id_user = json['id_user']
 
-    if (db.verify_register('lists', 'id', id)):
+    if (not db.verify_register('users', 'id', id_user)):
+        return jsonify(error(400, 'User not exists'))
+    elif (db.verify_register('lists', 'id', id)):
         return jsonify(error(400, 'List already exists'))
     else:
         # create instance 
@@ -127,31 +134,58 @@ def create_list():
             return jsonify(list_user.to_dict())
         else:
             return jsonify(error(203, 'Not Allowed'))
-   
+
+@app.route('/lists/<id_list>', methods=['POST'])
+def update_list(id_list):
+    if request.method == 'POST':
+        # query for check username 
+        query_db = db.read('lists', 'id', id_list)
+        list_user = query_db[0] if len(query_db) == 1 else None 
+        if(list_user == None):
+            return jsonify(error(404, 'List not exists'))
+        
+        else:
+            json = request.get_json()
+
+            id = list_user['id']
+            title = json['title']
+            date = list_user['date']
+            id_user = list_user['id_user']
+            link = list_user["link"]
+            
+            list_user = List(id, id_user, title, date)
+
+            if(db.update('lists', id_list, list_user.to_dict())):
+                return jsonify(list_user.to_dict())
+            
+            else:
+                return jsonify(error(203, 'Not Allowed'))
+    
 @app.route('/lists/delete/<id_list>', methods=['DELETE'])
 def delete_list(id_list):
-    # query for check list 
-    query_db = db.read('lists', 'id', id_list)
-    list_user = query_db[0] if len(query_db) == 1 else None 
-    if (list_user == None):
-        return jsonify(error(404, 'List not exists'))
-    
-    else:
-        if (db.delete('lists', id_list)):
-            
-            # delete all tasks 
-            tasks = db.read('tasks', 'id_list', id_list)
-            if (len(tasks) > 0):
-                for task in tasks:
-                    db.delete('tasks', task['id'])
-
-            return jsonify(error(200, "Removed"))
-
+    if request.method == 'DELETE':
+        # query for check list 
+        query_db = db.read('lists', 'id', id_list)
+        list_user = query_db[0] if len(query_db) == 1 else None 
+        if (list_user == None):
+            return jsonify(error(404, 'List not exists'))
+        
         else:
-            return jsonify(error(400, "Not removed"))
+            if (db.delete('lists', id_list)):
+                
+                # delete all tasks 
+                tasks = db.read('tasks', 'id_list', id_list)
+                if (len(tasks) > 0):
+                    for task in tasks:
+                        db.delete('tasks', task['id'])
+
+                return jsonify(error(200, "Removed"))
+
+            else:
+                return jsonify(error(400, "Not removed"))
 
   
-
+# routes task 
 @app.route('/tasks/<id_task>', methods=['GET'])
 def get_task(id_task):
     return jsonify(db.read('tasks', 'id', id_task))
@@ -169,7 +203,9 @@ def create_task():
     description = json['description']
     id_list = json['id_list']
 
-    if (db.verify_register('task', 'id', id)):
+    if (not db.verify_register('lists', 'id', id_list)):
+            return jsonify(error(400, 'List not exists'))
+    elif (db.verify_register('tasks', 'id', id)):
         return jsonify(error(400, 'Task already exists'))
     else:
         # create instance 
@@ -181,21 +217,51 @@ def create_task():
         else:
             return jsonify(error(203, 'Not Allowed'))
 
+@app.route('/tasks/<id_task>', methods=['POST'])
+def update_task(id_task):
+
+    if request.method == 'POST':
+
+        # query for check id task 
+        query_db = db.read('tasks', 'id', id_task)
+        task = query_db[0] if len(query_db) == 1 else None 
+        if(task == None):
+            return jsonify(error(404, 'Task not exists'))
+        
+        else:
+            json = request.get_json()
+            
+            id = task['id']
+            title = json['title']
+            description = json['description']
+            date = task['date']
+            id_list = task['id_list']
+            status = json["status"]
+            
+            task = Task(id=id, id_list=id_list,  title=title, description=description, date=date, status=status)
+
+            if(db.update('tasks', id_task, task.to_dict())):
+                return jsonify(task.to_dict())
+            
+            else:
+                return jsonify(error(203, 'Not Allowed'))
+
 @app.route('/tasks/delete/<id_task>', methods=['DELETE'])
 def delete_task(id_task):
-    # query for check task 
-    query_db = db.read('tasks', 'id', id_task)
-    task = query_db[0] if len(query_db) == 1 else None 
-    if (task == None):
-        return jsonify(error(404, 'Task not exists'))
-    
-    else:
-        if (db.delete('tasks', id_task)):
-
-            return jsonify(error(200, "Removed"))
-
+    if request.method == 'DELETE':
+        # query for check task 
+        query_db = db.read('tasks', 'id', id_task)
+        task = query_db[0] if len(query_db) == 1 else None 
+        if (task == None):
+            return jsonify(error(404, 'Task not exists'))
+        
         else:
-            return jsonify(error(400, "Not removed"))
+            if (db.delete('tasks', id_task)):
+
+                return jsonify(error(200, "Removed"))
+
+            else:
+                return jsonify(error(400, "Not removed"))
 
 if __name__ == '__main__':
     port = os.environ.get('PORT', 5000)
